Issue #14 - Retirada do handle genérico para Throwable
