+++ src/main/java/br/com/sicredi/canaisdigitais/avaliacaotecnicacanais/infra/exception/dto/ValidationError.java
@@ -1,0 +1,8 @@
+package br.com.sicredi.canaisdigitais.avaliacaotecnicacanais.infra.exception.dto;
+
+public record ValidationError(
+        String field,
+        String message
+) {
+
+}
