--- src/main/java/br/com/sicredi/canaisdigitais/avaliacaotecnicacanais/infra/exception/response/ValidationError.java
@@ -1,8 +1,0 @@
-package br.com.sicredi.canaisdigitais.avaliacaotecnicacanais.infra.exception.response;
-
-public record ValidationError(
-        String field,
-        String message
-) {
-
-}
